Move architecture-specific code to `src/math/arch`
Move the code and call into its new location with
`select_implementation`.
